修复ide文件名修改时带来的bug
