+++ public/mix-manifest.json
@@ -1,5 +1,5 @@
 {
-    "/vendor/js//vendor/js/app.chunk.js": "/vendor/js//vendor/js/app.chunk.js",
-    "/vendor/js/manifest.js": "/vendor/js/manifest.js",
-    "/vendor/js//vendor/js/vendor.chunk.js": "/vendor/js//vendor/js/vendor.chunk.js"
+    "/vendor/js//vendor/js/app.chunk.js": "/vendor/js//vendor/js/app.chunk.js?id=634d14e6b55b7968187c",
+    "/vendor/js/manifest.js": "/vendor/js/manifest.js?id=16328279cafb24d4ca21",
+    "/vendor/js//vendor/js/vendor.chunk.js": "/vendor/js//vendor/js/vendor.chunk.js?id=307f5ce8e9d2656ba1f8"
 }
