+++ resources/js/modules/views/authorities/authority/create.vue
@@ -1,14 +1,14 @@
 <template>
     <i-drawer title="权限添加" :width="720" :loading="loading">
-        <Form ref="formCreate" :model="create" :label-width="100" :rules="ruleValidate">
+        <Form ref="formCreate" :model="data" :label-width="100" :rules="ruleValidate">
             <FormItem label="权限名称" prop="name">
-                <Input v-model="create.name"></Input>
+                <Input v-model="data.name"></Input>
             </FormItem>
             <FormItem label="请求URI" prop="uri">
-                <Input v-model="create.uri"></Input>
+                <Input v-model="data.uri"></Input>
             </FormItem>
             <FormItem label="请求描述" prop="description">
-                <Input v-model="create.description" type="textarea" :rows="3"></Input>
+                <Input v-model="data.description" type="textarea" :rows="3"></Input>
             </FormItem>
             <FormItem label="分配菜单">
                 <div class="menu-box">
@@ -27,30 +27,13 @@
 <script>
     import contentDrawer from '../../../mixins/content-drawer'
     import IDrawer from "../../../components/content/drawer";
+    import Authority from './authority'
 
     export default {
         name: "create",
         components: {IDrawer},
-        mixins: [contentDrawer],
-        data() {
-            return {
-                loading: true,
-                create: {},
-                menus: {
-                    data: []
-                },
-                ruleValidate: {
-                    name: [
-                        {required: true, message: '权限名称必须填写', trigger: 'blur'},
-                        {type: 'string', min: 2, max: 20, message: '权限名称字符长度是2-20个字符', trigger: 'blur'}
-                    ],
-                    uri: [
-                        {required: true, message: '权限URI必须填写', trigger: 'blur'},
-                        {type: 'string', min: 2, max: 50, message: '权限URI字符长度是2-50个字符', trigger: 'blur'}
-                    ]
-                }
-            }
-        }, computed: {
+        mixins: [contentDrawer, Authority],
+        computed: {
             checkedMenus() {
                 return this.toChecked(JSON.parse(JSON.stringify(this.menus.data)))
             }
@@ -61,40 +44,16 @@
         }, methods: {
             submit(name) {
                 this.validate(name).then(() => {
+                    this.loading = true;
                     this.$http.post(`authorities/authority`,
-                        Object.assign({}, this.create, {menus: this.checkedMenus})
+                        Object.assign({}, this.data, {menus: this.checkedMenus})
                     ).then(() => {
                         this.closeDrawer(false)
+                    }).finally(() => {
+                        this.loading = false;
                     });
                 }).catch(() => {
                 });
-            },
-            setTreeData(source) {
-                let cloneData = JSON.parse(JSON.stringify(source))
-                let tree = cloneData.filter(father => {
-                    let branchArr = cloneData.filter(child => {
-                        return father['id'] == child['parent_id']
-                    });
-                    if (branchArr.length > 0) {
-                        father['children'] = branchArr
-                    }
-                    return father['parent_id'] == 0
-                })
-                return tree.map((item) => {
-                    return Object.assign(item, {expand: true});
-                })
-            },
-            toChecked(data) {
-                let arr = [];
-                data.forEach((item) => {
-                    if (item.indeterminate === true || item.checked === true) {
-                        item.checked === true && arr.push(item.id);
-                        if (item.children && item.children.length > 0) {
-                            arr = arr.concat(this.toChecked(item.children));
-                        }
-                    }
-                });
-                return arr
             }
         }
     }
