+++ resources/js/modules/views/authorities/authority/update.vue
@@ -1,14 +1,14 @@
 <template>
     <i-drawer title="权限添加" :loading="loading" :width="720">
-        <Form ref="formUpdate" :model="update" :label-width="100" :rules="ruleValidate">
+        <Form ref="formUpdate" :model="data" :label-width="100" :rules="ruleValidate">
             <FormItem label="权限名称" prop="name">
-                <Input v-model="update.name"></Input>
+                <Input v-model="data.name"></Input>
             </FormItem>
             <FormItem label="请求URI" prop="uri">
-                <Input v-model="update.uri"></Input>
+                <Input v-model="data.uri"></Input>
             </FormItem>
             <FormItem label="请求描述" prop="description">
-                <Input v-model="update.description" type="textarea" :rows="3"></Input>
+                <Input v-model="data.description" type="textarea" :rows="3"></Input>
             </FormItem>
             <FormItem label="分配菜单">
                 <div class="menu-box">
@@ -27,14 +27,15 @@
 <script>
     import contentDrawer from '../../../mixins/content-drawer'
     import IDrawer from "../../../components/content/drawer";
+    import Authority from './authority'
 
     export default {
         name: "update",
         components: {IDrawer},
-        mixins: [contentDrawer],
+        mixins: [contentDrawer, Authority],
         mounted() {
             this.$http.get(`authorities/authority/${this.props.id}/edit`).then((res) => {
-                this.update = res.data
+                this.data = res.data
                 this.menus.data = res.menus
                 let data = [];
                 JSON.parse(JSON.stringify(this.menus.data)).forEach((item) => {
@@ -50,25 +51,7 @@
                 this.menus.data = this.setTreeData(data)
             }).finally(() => this.loading = false);
         },
-        data() {
-            return {
-                loading: true,
-                update: {},
-                menus: {
-                    data: []
-                },
-                ruleValidate: {
-                    name: [
-                        {required: true, message: '权限名称必须填写', trigger: 'blur'},
-                        {type: 'string', min: 2, max: 20, message: '权限名称字符长度是2-20个字符', trigger: 'blur'}
-                    ],
-                    uri: [
-                        {required: true, message: '权限URI必须填写', trigger: 'blur'},
-                        {type: 'string', min: 2, max: 50, message: '权限URI字符长度是2-50个字符', trigger: 'blur'}
-                    ]
-                }
-            }
-        }, computed: {
+        computed: {
             checkedMenus() {
                 return this.toChecked(JSON.parse(JSON.stringify(this.menus.data)))
             }
@@ -76,65 +59,16 @@
         methods: {
             submit(name) {
                 this.validate(name).then(() => {
+                    this.loading = true;
                     this.$http.put(`authorities/authority/${this.props.id}`,
-                        Object.assign({}, this.update, {menus: this.checkedMenus})
+                        Object.assign({}, this.data, {menus: this.checkedMenus})
                     ).then(() => {
                         this.closeDrawer(false)
+                    }).finally(() => {
+                        this.loading = false;
                     });
                 }).catch(() => {
                 });
-            },
-            setTreeData(source) {
-                let cloneData = JSON.parse(JSON.stringify(source))
-                let tree = cloneData.filter(father => {
-                    let branchArr = cloneData.filter(child => {
-                        return father['id'] == child['parent_id']
-                    });
-                    if (branchArr.length > 0) {
-                        father['children'] = branchArr
-                    }
-                    return father['parent_id'] == 0
-                })
-                return tree.map((item) => {
-                    return Object.assign(item, {expand: true});
-                })
-            },
-            child(parent) {
-                return this.menus.data.filter(val => val.parent_id == parent)
-            },
-            indeterminate(item) {
-                let hasChild = this.child(item.id).length > 0
-                if (!hasChild) {
-                    return false
-                }
-                let has = typeof this.update.menus.find(val => val === item.id) !== 'undefined'
-                if (has) {
-                    return true
-                }
-                return false;
-            },
-            checked(item) {
-                let hasChild = this.child(item.id).length > 0
-                if (hasChild) {
-                    return false
-                }
-                let has = typeof this.update.menus.find(val => val === item.id) !== 'undefined'
-                if (has) {
-                    return true
-                }
-                return false;
-            },
-            toChecked(data) {
-                let arr = [];
-                data.forEach((item) => {
-                    if (item.indeterminate === true || item.checked === true) {
-                        item.checked === true && arr.push(item.id);
-                        if (item.children && item.children.length > 0) {
-                            arr = arr.concat(this.toChecked(item.children));
-                        }
-                    }
-                });
-                return arr
             }
         }
     }
