+++ resources/js/modules/views/authorities/menu/update.vue
@@ -1,27 +1,27 @@
 <template>
     <i-drawer title="菜单修改" :width="720" :loading="loading">
-        <Form ref="formUpdate" :model="update" :label-width="100" :rules="ruleValidate">
+        <Form ref="formUpdate" :model="data" :label-width="100" :rules="ruleValidate">
             <FormItem label="上级菜单">
-                <Input v-model="update.parent.name" disabled></Input>
+                <Input v-model="data.parent.name" disabled></Input>
             </FormItem>
             <FormItem label="菜单名称" prop="name">
-                <Input v-model="update.name"></Input>
+                <Input v-model="data.name"></Input>
             </FormItem>
             <FormItem label="菜单url" prop="url">
-                <Input v-model="update.url"></Input>
+                <Input v-model="data.url"></Input>
             </FormItem>
             <FormItem label="菜单描述" prop="description">
-                <Input v-model="update.description" type="textarea"></Input>
+                <Input v-model="data.description" type="textarea"></Input>
             </FormItem>
             <FormItem label="排序" prop="sort">
-                <Input v-model="update.sort" number></Input>
+                <Input v-model="data.sort" number></Input>
             </FormItem>
             <FormItem label="分配权限">
                 <Transfer
                         :titles="['可分配权限', '已有权限']"
                         :list-style="{width: '250px',height: '500px'}"
                         :data="authorities.data"
-                        :target-keys="update.authorities"
+                        :target-keys="data.authorities"
                         @on-change="handleChange"></Transfer>
             </FormItem>
         </Form>
@@ -34,56 +34,35 @@
 <script>
     import IDrawer from "../../../components/content/drawer";
     import contentDrawer from '../../../mixins/content-drawer'
+    import Menu from './menu';
+
 
     export default {
         name: "update",
-        mixins: [contentDrawer],
+        mixins: [contentDrawer, Menu],
         components: {IDrawer},
         mounted(){
             this.$http.get(`authorities/menu/${this.props.id}/edit`).then((res) => {
                 res.data.parent = res.data.parent || {name: '顶级菜单'};
-                this.update = res.data
+                this.data = res.data
                 this.authorities.data = res.authorities
             }).finally(() => this.loading = false);
-        },
-        data(){
-            return {
-                loading: true,
-                update: {
-                    parent: {},
-                    authorities: []
-                },
-                authorities: {
-                    data: []
-                },
-                ruleValidate: {
-                    name: [
-                        {required: true, message: '菜单名称必须填写', trigger: 'blur'},
-                        {type: 'string', min: 2, max: 20, message: '菜单名称字符长度是2-20个字符', trigger: 'blur'}
-                    ],
-                    url: [
-                        {required: true, message: '菜单url必须填写', trigger: 'blur'},
-                        {type: 'string', min: 2, max: 50, message: '菜单url字符长度是2-50个字符', trigger: 'blur'},
-                    ],
-                    sort: [
-                        {required: true, type: 'number', message: '排序必须填写', trigger: 'blur'},
-                        {type: 'number', min: 0, max: 99, message: '排序数值在0-99之间', trigger: 'blur'},
-                    ]
-                }
-            }
         },
         methods: {
             submit(name) {
                 this.validate(name).then(() => {
-                    this.$http.put(`authorities/menu/${this.props.id}`, this.update).then(() => {
+                    this.loading = true;
+                    this.$http.put(`authorities/menu/${this.props.id}`, this.data).then(() => {
                         this.closeDrawer(false)
+                    }).finally(() => {
+                        this.loading = false;
                     });
                 }).catch(() => {
 
                 });
             },
             handleChange(newTargetKeys) {
-                this.update.authorities = newTargetKeys
+                this.data.authorities = newTargetKeys
             }
         }
     }
