+++ resources/js/modules/views/authorities/role/create.vue
@@ -4,13 +4,13 @@
             <Step title="分配菜单"></Step>
             <Step title="分配权限"></Step>
         </Steps>
-        <Form :model="create" :label-width="100" :rules="ruleValidate" ref="formCreate">
+        <Form :model="data" :label-width="100" :rules="ruleValidate" ref="formCreate">
             <div v-show="current === 0">
                 <FormItem label="部门名称" prop="name">
-                    <Input v-model="create.name"></Input>
+                    <Input v-model="data.name"></Input>
                 </FormItem>
                 <FormItem label="部门描述" prop="description">
-                    <Input v-model="create.description" type="textarea" :rows="6"></Input>
+                    <Input v-model="data.description" type="textarea" :rows="6"></Input>
                 </FormItem>
                 <FormItem label="分配菜单">
                     <div class="menu-box">
@@ -26,18 +26,18 @@
                             :titles="['可分配权限', '已有权限']"
                             :list-style="{width: '250px',height: '500px'}"
                             :data="authorities.data"
-                            :target-keys="create.authorities"
+                            :target-keys="data.authorities"
                             @on-change="handleChange"></Transfer>
                 </FormItem>
             </div>
         </Form>
         <div slot="footer">
-            <Button type="primary" v-if="current === 1" @click="next">
+            <Button type="primary" v-if="current === 1" @click="next('formCreate')">
                 <Icon type="ios-arrow-back"></Icon>
                 上一步
             </Button>
             <Button type="primary" v-if="current === 1" icon="ios-add" @click="submit('formCreate')">提交</Button>
-            <Button type="primary" v-if="current === 0" @click="next">下一步
+            <Button type="primary" v-if="current === 0" @click="next('formCreate')">下一步
                 <Icon type="ios-arrow-forward"></Icon>
             </Button>
         </div>
@@ -47,11 +47,12 @@
 <script>
     import contentDrawer from '../../../mixins/content-drawer'
     import IDrawer from "../../../components/content/drawer";
+    import Role from './role';
 
     export default {
         name: "create",
         components: {IDrawer},
-        mixins: [contentDrawer],
+        mixins: [contentDrawer, Role],
         mounted() {
             this.$http.get(`authorities/role/create`).then((res) => {
                 this.menus.data = this.setTreeData(res)
@@ -59,68 +60,22 @@
                 this.loading = false
             });
         },
-        data() {
-            return {
-                current: 0,
-                loading: true,
-                create: {
-                    authorities: [],
-                    menus: []
-                },
-                authorities: {
-                    data: []
-                },
-                menus: {
-                    data: []
-                },
-                ruleValidate: {
-                    name: [
-                        {required: true, message: '部门名称不能为空', trigger: 'blur'},
-                        {type: 'string', min: 2, max: 20, message: '用户名称字符长度是2-20个字符', trigger: 'blur'}
-                    ],
-                    description: [
-                        {type: 'string', max: 255, message: '部门描述最长255个字符', trigger: 'blur'}
-                    ]
-                },
-                checked: false
-            }
-        },
         computed: {
             checkedMenus() {
                 return this.toChecked(JSON.parse(JSON.stringify(this.menus.data)))
             }
         },
         methods: {
-            handleChange(newTargetKeys) {
-                this.create.authorities = newTargetKeys
-            },
+
             submit(name) {
-                this.validate('formCreate').then(() => {
-                    this.$http.post(`authorities/role`, this.create).then(() => {
+                this.validate(name).then(() => {
+                    this.loading = true;
+                    this.$http.post(`authorities/role`, this.data).then(() => {
                         this.closeDrawer(false)
+                    }).finally(() => {
+                        this.loading = false;
                     });
                 }).catch();
-            },
-            getAuthorities(){
-                this.$http.get(`authorities/menu/authority`, {
-                    params: {
-                        ids: this.checkedMenus
-                    }
-                }).then((res) => {
-                    this.authorities.data = this.toTransfer(res)
-                });
-            },
-            next() {
-                if (this.current === 0) {
-                    this.validate('formCreate').then(() => {
-                        if (this.checked === true) {
-                            this.getAuthorities();
-                        }
-                        this.current = ++this.current
-                    }).catch();
-                } else {
-                    this.current = --this.current;
-                }
             },
             setTreeData(source) {
                 let cloneData = JSON.parse(JSON.stringify(source))
@@ -148,25 +103,11 @@
                     }
                 });
                 return arr
-            },
-            toTransfer(data) {
-                let lists = [];
-                data.forEach((item) => {
-                    item.authorities.forEach((val) => {
-                        if (lists.findIndex((v) => v.key === val.id) === -1) {
-                            lists.push({
-                                key: val.id,
-                                label: `${val.name}`
-                            })
-                        }
-                    })
-                })
-                return lists
             }
         },
         watch: {
             checkedMenus(val) {
-                this.create.menus = val
+                this.data.menus = val
                 this.checked = true
             }
         }
