+++ webpack.mix.js
@@ -7,7 +7,7 @@ mix.extract(['vue', 'axios']);
 mix.webpackConfig({
     output: {
         filename: '[name].js',
-        chunkFilename: mix.inProduction() ? 'vendor/js/[name].[chunkhash].chunk.js' : 'vendor/js/[name].chunk.js'
+        chunkFilename: mix.inProduction() ? 'vendor/js/[name].chunk.js' : 'vendor/js/[name].chunk.js'
     }
 });
 mix.options({
